\name{UncertWebRTools-package}
\alias{UncertWebRTools-package}
\alias{UncertWeb}
\docType{package}
\title{UncertWeb}
\description{By now, reading and writing some UNetCDF files}
\details{
\tabular{ll}{
Package: \tab UncertWeb\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-29\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{Benedikt Gr\"aler \email{ben.graeler@uni-muenster.de}}
\references{UncertWeb wiki}
\keyword{package, UncertML}
% \usage{Currently there are two functions reading and writing some UNetCDF files}
% \seealso{ }
% \examples{ }
