\name{UncertWeb-package}
\alias{UncertWeb-package}
\alias{UncertWeb}
\docType{package}
\title{UncertWeb}
\description{By now, reading and writing some UNetCDF files}
\details{
\tabular{ll}{
Package: \tab UncertWeb\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-21\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{Benedikt Gräler \email{ben.graeler@uni-muenster.de}}
\references{UncertWeb wiki}
\keyword{package, UncertML}
\usage{Currently there are two functions reading and writing some UNetCDF files}
\seealso{ }
\examples{ }
