\name{JsonXmlUncertWebRTools-package}
\alias{JsonXmlUncertWebRTools-package}
\alias{UncertWeb}
\docType{package}
\title{UncertWeb}
\description{Reading and writing XML / JSON Uncertainty types, and some sampling from distributions using existing R functions}
\details{
\tabular{ll}{
Package: \tab JsonXmlUncertWebRTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-29\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{Lucy Bastin \email{l.bastin@aston.ac.uk}}
\references{UncertWeb wiki}
\keyword{package, UncertML}
% \usage{Reading and writing XML / JSON Uncertainty types, and some sampling from distributions using existing R functions}
% \seealso{ }
% \examples{ }
